Revert "app engine connect sql"
This reverts commit 176306d1f65285d6679846bbe48050494e6778ff [formerly 8b73d00420a3c8d692eef420b1b0f01e80e49342].


Former-commit-id: d79afd22d792b3436729198531278a87c774d1ef

+++ gin-project/dao/model.go
@@ -1,6 +1,6 @@
 package dao
 
 type User struct {
-	ID   int64  `gorm:"primary_key"`
+	ID   uint   `gorm:"primary_key"`
 	Name string `gorm:"size:255"` // string默认长度为255, 使用这种tag重设。
 }
